CHANGE: バグ修正, History を Standalone で使えるように変更, step を csv に追加

+++ v1/variable_manager.py
@@ -116,4 +116,4 @@ class VariableManager:
             return {name: var.value for name, var in raw.items()}
 
         elif format == 'values':
-            return np.array([var.value for var in raw.values()], dtype=np.float64)
+            return np.array([var.value for var in raw.values()])
